+++ swift/Tests/LibSignalClientTests/NativeTestingNice.swift
@@ -1439,6 +1439,16 @@ internal enum NativeTestingNice {
         >.convertReturn(consuming: rawOutput)
 
     }
+    internal static func TESTING_SetPushTokenApnsTests() throws -> [GrpcTestCase<String, Void>] {
+        var rawOutput = GrpcTestCaseVecConverter<StringConverter, VoidConverter>.emptyFfiReturn()
+        try checkError(
+            SignalFfi.signal_testing_set_push_token_apns_tests(
+                &rawOutput,
+            )
+        )
+        return try GrpcTestCaseVecConverter<StringConverter, VoidConverter>.convertReturn(consuming: rawOutput)
+
+    }
     internal static func TESTING_SetUsernameLinkTests() throws -> [GrpcTestCase<
         SetUsernameLinkArgs, SetUsernameLinkOut
     >] {